liczby rzeczywiste zamiast liczb naturalnych

+++ Prostokat/Prostokat/Program.cs
@@ -6,12 +6,12 @@ namespace Prostokat
     {
         static void Main(string[] args)
         {
-            int side_a;
-            int side_b;
+            double side_a;
+            double side_b;
 
             Console.WriteLine("Podaj bok A (cm):");
             string user_input_side_a = Console.ReadLine();
-            if (int.TryParse(user_input_side_a, out side_a) == false)
+            if (double.TryParse(user_input_side_a, out side_a) == false)
             {
                 Console.WriteLine("Błędne dane podane z klawiatury");
                 return;
@@ -19,13 +19,13 @@ namespace Prostokat
 
             Console.WriteLine("Podaj bok B (cm):");
             string user_input_side_b = Console.ReadLine();
-            if (int.TryParse(user_input_side_b, out side_b) == false)
+            if (double.TryParse(user_input_side_b, out side_b) == false)
             {
                 Console.WriteLine("Błędne dane podane z klawiatury");
                 return;
             }
 
-            int field = side_a * side_b;
+            double field = side_a * side_b;
 
             Console.WriteLine("Pole prostokąta wynosi: {0}cm/kw", field);
             Console.ReadKey();
